Make accessor functions of particles inline
