+++ source/particles/CMakeLists.txt
@@ -19,8 +19,6 @@ INCLUDE_DIRECTORIES(BEFORE ${CMAKE_CURRENT_BINARY_DIR})
 SET(_src
   data_out.cc
   particle.cc
-  particle_accessor.cc
-  particle_iterator.cc
   particle_handler.cc
   generators.cc
   property_pool.cc
@@ -30,8 +28,6 @@ SET(_src
 SET(_inst
   data_out.inst.in
   particle.inst.in
-  particle_accessor.inst.in
-  particle_iterator.inst.in
   particle_handler.inst.in
   generators.inst.in
   utilities.inst.in
